Improved the standard implementation of the regeneration driver.
Do not expose the default implementation for the regeneration driver
but provide a function that returns a pointer to the default implementation.
Moreover, do not free the regeneration driver on exit. This is not needed
since the default is now a global constant implementation and the user provided
one must be freed by the caller.